requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
anyio==4.7.0
APScheduler==3.6.3
astunparse==1.6.3
attrs==24.2.0
backoff==2.2.1
cachetools==4.2.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
commonmark==0.9.1
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
docker==7.1.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.55.2
frozenlist==1.5.0
fsspec==2024.10.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inference-cli==0.29.2
Jinja2==3.1.4
keras==3.7.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.3
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.4.2
numpy==1.26.4
nvidia-ml-py==12.560.30
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
pillow==10.4.0
propcache==0.2.1
protobuf==5.29.0
psutil==6.1.0
py-cpuinfo==9.0.0
pybase64==1.0.2
Pygments==2.18.0
pyparsing==3.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-telegram-bot==21.9
pytz==2024.2
pywin32==308
PyYAML==6.0.2
requests==2.32.3
rich==13.0.1
scipy==1.14.1
seaborn==0.13.2
setuptools==75.6.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
supervision==0.22.0
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tk==0.1.0
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ultralytics==8.3.47
ultralytics-thop==2.0.12
urllib3==2.2.3
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
yarl==1.18.3
